requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-groq>=0.1.0
langchain-google-genai>=0.0.10
langchain-sambanova>=0.1.0
langchain-community>=0.1.0
langchain-huggingface>=0.0.10
playwright>=1.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0
chroma-db>=0.4.0
langgraph>=0.1.0
nest-asyncio>=1.5.0
